+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # In progress
 - Fix green border for some android versions
 - Display html entity descriptions
+- Enable link input without camera permission for qr scan
 
 # 0.8.15
 - Fix poll option padding
@@ -10,7 +11,7 @@
 - Fix hidden process bug
 
 # 0.8.14
-- Fix network initialization error
+- Fix network initialization erro
 - Fix notification linking
 - Enable in-app notifications