+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# In progress
+# 0.8.16
 - Fix green border for some android versions
 - Display html entity descriptions
 - Enable link input without camera permission for qr scan
 - Enable ipfs image urls
 - Implement app telemetry
+- Fix bootnode url changing
 
 # 0.8.15
 - Fix poll option padding